# [julia] Refactor julia installation, and benchmark code