--- a/julia_con/RelatedCon/src/RelatedCon.jl
+++ b/julia_con/RelatedCon/src/RelatedCon.jl
@@ -4,6 +4,7 @@ using JSON3
 using StructTypes
 using Dates
 using PrecompileTools
+using Base.Threads: @threads
 
 const topn = 5