--- a/julia_con/related.jl
+++ b/julia_con/related.jl
@@ -3,7 +3,6 @@ using Base.Threads
 using JSON3
 using StructTypes
 using Dates
-using StaticArrays
 using ChunkSplitters
 
 # warmup is done by hyperfine