~~This design uses  a threadsafe task-local-buffer to pre-allocate maxn and maxv on each spawned task. This obviates the need for static arrays and does buffer-reuse instead.~~

~~On my machine I see the runtime of the single threaded code drop from 23ms to 17ms, and the concurrent benchmark drops from 5ms to 3ms. I'll update with better benchmarks once Docker finally finishes setting up on my machine.~~

This refactors the code, changing how Julia is installed to use the official binaries instead of the ones that are incorrectly linked to LLVM, and it bundles the code into a package to use native precompilation instead of manually warming up the analysis functions. This makes it more reproducible, and more in line with how the AOT compiled languages operate. 

There's no significant speed difference that I see locally.