@MasonProtter 2023-10-12
Hmm, that's weird. When I run the Docker benchmarks, I'm seeing a performance regression relative to `main` rather than a performance improvement. Not sure what's up with that, I suspect I'm just doing something wrong with Docker.