@jinyus 2023-10-12
Performance is the same on my machine will test on the benchmarking VM