@jinyus 2023-10-12
These are the results from the VM:

```
Julia:

        Benchmark 1: julia related.jl
        Processing time (w/o IO): 38 milliseconds
        Processing time (w/o IO): 38 milliseconds
        Processing time (w/o IO): 38 milliseconds
        Processing time (w/o IO): 38 milliseconds
        Processing time (w/o IO): 30 milliseconds
        Processing time (w/o IO): 31 milliseconds
        Processing time (w/o IO): 30 milliseconds
        Processing time (w/o IO): 38 milliseconds
        Processing time (w/o IO): 31 milliseconds
        Processing time (w/o IO): 31 milliseconds
        Processing time (w/o IO): 31 milliseconds
        Processing time (w/o IO): 38 milliseconds
        Processing time (w/o IO): 37 milliseconds
          Time (mean ± σ):      2.762 s ±  0.022 s    [User: 2.585 s, System: 0.176 s]
          Range (min … max):    2.730 s …  2.802 s    10 runs
         
Julia:

        Benchmark 1: julia related.jl
        Processing time (w/o IO): 461 milliseconds
        Processing time (w/o IO): 575 milliseconds
        Processing time (w/o IO): 459 milliseconds
          Time (mean ± σ):      3.852 s ±  0.141 s    [User: 3.643 s, System: 0.207 s]
          Range (min … max):    3.752 s …  3.951 s    2 runs
         
Julia:

        Benchmark 1: julia related.jl
        Processing time (w/o IO): 5052 milliseconds
        Processing time (w/o IO): 5046 milliseconds
        Processing time (w/o IO): 5046 milliseconds
          Time (mean ± σ):     13.294 s ±  0.013 s    [User: 12.972 s, System: 0.311 s]
          Range (min … max):   13.285 s … 13.304 s    2 runs

```