@MasonProtter 2023-10-12
Hm that's really interesting that there's such a big difference between the two versions for me when i just invoke `julia related.jl` from the command line. 

I wonder if this has to do with the way julia was installed on the docker system. You installed it using `pacman`, but that's explicitly not recommended because the `pacman` version is compiled against the system LLVM which does not yet have all the bug fixes and patches we are working on upstreaming. See the arch wiki here: https://wiki.archlinux.org/title/Julia:
```quote
Note: The [julia](https://archlinux.org/packages/?name=julia) package is compiled against system libraries and is therefore unsupported by the upstream. While it may be suitable for basic usage it is incompatible with external packages (e.g. MKL, LLVM, LLVMExtra, or packages that depend on these, like CUDA or Flux) which rely on Julia's bundled versions of its dependencies. For example, Sundials.jl (which is a dependency of DifferentialEquations.jl) [fails](https://github.com/SciML/Sundials.jl/issues/358) to precompile with this package, but builds without errors with the official builds. Bugs reported against this binary will be [closed](https://github.com/JuliaLinearAlgebra/MKL.jl/issues/105#issuecomment-1025985667).
```