@MasonProtter 2023-10-13
@jinyus, would you mind re-running the benchmark. I want to check if what you see matches what I'm seeing here.