@jinyus 2023-10-13
Results Run 1:

```
Julia:

        Benchmark 1: julia related.jl
        Processing time (w/o IO): 32 milliseconds
        Processing time (w/o IO): 31 milliseconds
        Processing time (w/o IO): 32 milliseconds
        Processing time (w/o IO): 39 milliseconds
        Processing time (w/o IO): 39 milliseconds
        Processing time (w/o IO): 39 milliseconds
        Processing time (w/o IO): 31 milliseconds
        Processing time (w/o IO): 39 milliseconds
        Processing time (w/o IO): 31 milliseconds
        Processing time (w/o IO): 38 milliseconds
        Processing time (w/o IO): 38 milliseconds
        Processing time (w/o IO): 38 milliseconds
        Processing time (w/o IO): 39 milliseconds
          Time (mean ± σ):      2.725 s ±  0.024 s    [User: 2.617 s, System: 0.212 s]
          Range (min … max):    2.685 s …  2.761 s    10 runs
         
Julia:

        Benchmark 1: julia related.jl
        Processing time (w/o IO): 576 milliseconds
        Processing time (w/o IO): 459 milliseconds
        Processing time (w/o IO): 462 milliseconds
          Time (mean ± σ):      3.709 s ±  0.016 s    [User: 3.556 s, System: 0.255 s]
          Range (min … max):    3.697 s …  3.720 s    2 runs
         
Julia:

        Benchmark 1: julia related.jl
        Processing time (w/o IO): 3993 milliseconds
        Processing time (w/o IO): 5056 milliseconds
        Processing time (w/o IO): 4002 milliseconds
          Time (mean ± σ):     12.172 s ±  1.471 s    [User: 11.938 s, System: 0.330 s]
          Range (min … max):   11.132 s … 13.212 s    2 runs
         
```

Run 2:

```
Julia:

        Benchmark 1: julia related.jl
        Processing time (w/o IO): 32 milliseconds
        Processing time (w/o IO): 38 milliseconds
        Processing time (w/o IO): 38 milliseconds
        Processing time (w/o IO): 31 milliseconds
        Processing time (w/o IO): 38 milliseconds
        Processing time (w/o IO): 38 milliseconds
        Processing time (w/o IO): 39 milliseconds
        Processing time (w/o IO): 39 milliseconds
        Processing time (w/o IO): 31 milliseconds
        Processing time (w/o IO): 31 milliseconds
        Processing time (w/o IO): 39 milliseconds
        Processing time (w/o IO): 38 milliseconds
        Processing time (w/o IO): 39 milliseconds
          Time (mean ± σ):      2.703 s ±  0.023 s    [User: 2.600 s, System: 0.206 s]
          Range (min … max):    2.666 s …  2.748 s    10 runs
         
Julia:

        Benchmark 1: julia related.jl
        Processing time (w/o IO): 463 milliseconds
        Processing time (w/o IO): 577 milliseconds
        Processing time (w/o IO): 462 milliseconds
          Time (mean ± σ):      3.885 s ±  0.151 s    [User: 3.717 s, System: 0.270 s]
          Range (min … max):    3.778 s …  3.991 s    2 runs
         
Julia:

        Benchmark 1: julia related.jl
        Processing time (w/o IO): 5052 milliseconds
        Processing time (w/o IO): 3999 milliseconds
        Processing time (w/o IO): 5052 milliseconds
          Time (mean ± σ):     12.195 s ±  1.472 s    [User: 11.984 s, System: 0.308 s]
          Range (min … max):   11.154 s … 13.236 s    2 runs

```